FIX: Compatibility with numpy 1.13

+++ bluesky/callbacks/scientific.py
@@ -101,7 +101,7 @@ class PeakStats(CollectThenCompute):
         self.min = x[np.argmin(y)], self.y_data[np.argmin(y)],
         self.com, = np.interp(center_of_mass(y), np.arange(len(x)), x)
         mid = (np.max(y) + np.min(y)) / 2
-        crossings = np.where(np.diff(y > mid))[0]
+        crossings = np.where(np.diff((y > mid).astype(np.int)))[0]
         _cen_list = []
         for cr in crossings.ravel():
             _x = x[cr:cr+2]
